Add documentation and cleanup example
