+++ examples/freertos/xlink/src/xlink_rx/xlink_rx.h
@@ -1,0 +1,9 @@
+// Copyright 2022 XMOS LIMITED.
+// This Software is subject to the terms of the XMOS Public Licence: Version 1.
+
+#ifndef XLINK_RX_H_
+#define XLINK_RX_H_
+
+void create_xlink_rx_tasks(unsigned priority);
+
+#endif /* XLINK_RX_H_ */
