+++ examples/freertos/xlink/src/xlink_tx/xlink_tx.h
@@ -1,0 +1,9 @@
+// Copyright 2022 XMOS LIMITED.
+// This Software is subject to the terms of the XMOS Public Licence: Version 1.
+
+#ifndef XLINK_TX_H_
+#define XLINK_TX_H_
+
+void create_xlink_tx_tasks(unsigned priority);
+
+#endif /* XLINK_TX_H_ */
