+++ examples/freertos/xlink/xlink.cmake
@@ -2,7 +2,12 @@
 # Gather Sources
 #**********************
 file(GLOB_RECURSE APP_SOURCES ${CMAKE_CURRENT_LIST_DIR}/src/*.c)
-set(APP_INCLUDES ${CMAKE_CURRENT_LIST_DIR}/src)
+set(APP_INCLUDES
+    ${CMAKE_CURRENT_LIST_DIR}/src
+    ${CMAKE_CURRENT_LIST_DIR}/src/link
+    ${CMAKE_CURRENT_LIST_DIR}/src/xlink_rx
+    ${CMAKE_CURRENT_LIST_DIR}/src/xlink_tx
+)
 add_subdirectory(${CMAKE_CURRENT_LIST_DIR}/bsp_config)
 
 #**********************
